podman/ultralytics/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
kafka-python>=2.0.2
prometheus-client>=0.17.0
loguru>=0.7.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
timm>=0.9.0
albumentations>=1.3.0 